Revert "Pull request antes da previa"
