+++ Cali_Imperium/Cali_Imperium/Sistema/Atualizar perfil/Atualizar perfil.aspx
@@ -1,0 +1,16 @@
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeBehind="Atualizar perfil.aspx.cs" Inherits="Cali_Imperium.Atualizar_perfil" %>
+
+<!DOCTYPE html>
+
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head runat="server">
+<meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
+    <title></title>
+</head>
+<body>
+    <form id="form1" runat="server">
+        <div>
+        </div>
+    </form>
+</body>
+</html>
