+++ Cali_Imperium/Cali_Imperium/Sistema/Perfil/css/atualizar.css
@@ -1,0 +1,104 @@
+﻿* {
+    margin: 0;
+    padding: 0;
+}
+
+body {
+    font: 100% normal Arial, Helvetica, sans-serif;
+    background: #161712;
+}
+
+form, input, select {
+    margin: 0;
+    padding: 0;
+    color: #ffffff;
+}
+
+div.logar {
+    background: #222222;
+    position: relative;
+    top: 50px;
+    border: 1px solid #262626;
+    width: 40%;
+    margin: 0 auto;
+    padding: 10%;
+    border-radius: 10px;
+}
+
+    div.logar h1 {
+        color: #a6a2ad;
+        font-size: 15px;
+        text-transform: uppercase;
+        padding: 5px 0 5px 5px;
+        text-align: center;
+        font-size: 14px;
+        margin-bottom: 30px;
+        display: block;
+    }
+
+
+
+    div.logar .atual {
+        padding: 10px 10px;
+        background: #262626;
+        border-bottom: 1px double #171717;
+        border-top: 1px double #171717;
+        border-left: 1px double #333333;
+        border-right: 1px double #333333;
+        position: relative;
+        display: block;
+        width: 85%;
+        height: 20px;
+        margin-bottom: 20px;
+        border-radius: 5px;
+    }
+
+button.btnCancelar {
+    width: 38%;
+    height: 35px;
+    background-color: rgb(255, 82, 29);
+    color: #fff;
+    margin-left: 15px;
+    margin-right: 15px;
+    border-radius: 3px;
+    border: solid;
+    border-color: rgb(255, 82, 29);
+    border-width: 1px;
+    cursor: pointer;
+    transition: 0.3s;
+    outline: none;
+    margin-left: 50px;
+}
+
+button.btnAlterar {
+    width: 40%;
+    height: 35px;
+    background-color: rgb(72, 146, 124);
+    color: #fff;
+    margin-right: 15px;
+    border-radius: 3px;
+    border: solid;
+    border-color: rgb(72, 146, 124);
+    border-width: 1px;
+    cursor: pointer;
+    transition: 0.3s;
+    outline: none;
+}
+
+select {
+    padding: 10px 10px;
+    width: 200px;
+    background: #262626;
+    border-bottom: 1px double #171717;
+    border-top: 1px double #171717;
+    border-left: 1px double #333333;
+    border-right: 1px double #333333;
+    width: 89%;
+    height: 43px;
+    margin-bottom: 20px;
+}
+
+.pp {
+    margin-left: 50px;
+    align-self: center;
+}
