+++ Cali_Imperium/Cali_Imperium/Sistema/Temperatura.cs
@@ -150,20 +150,20 @@ namespace Caliimperium
             }
         }
 
-        public int Pegar1Quartil(int id)
+        public int Pegar2Quartil(int id)
         {
             using (SqlConnection conn = new SqlConnection("Server=tcp:controlterm.database.windows.net,1433;Initial Catalog=ControlTerm;Persist Security Info=False;User ID=Control;Password=Term2k18;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"))
             {
                 conn.Open();
 
-                using (SqlCommand cmd = new SqlCommand("SELECT DISTINCT PERCENTILE_CONT(0.25) WITHIN GROUP(ORDER BY temperatura) OVER(PARTITION BY 1) as '1q' from temperatura where codArduino = (SELECT codArduino FROM Arduino where codUsuario = @id)", conn))
+                using (SqlCommand cmd = new SqlCommand("SELECT DISTINCT PERCENTILE_CONT(0.25) WITHIN GROUP(ORDER BY temperatura) OVER(PARTITION BY 1) as '2q' from temperatura where codArduino = (SELECT codArduino FROM Arduino where codUsuario = @id)", conn))
                 {
                     cmd.Parameters.AddWithValue("@id", id);
                     using (SqlDataReader dr = cmd.ExecuteReader())
                     {
                         if (dr.Read() == true)
                         {
-                            return int.Parse(dr["1q"].ToString()); // retorna a temperatura que esta no banco
+                            return int.Parse(dr["2q"].ToString()); // retorna a temperatura que esta no banco
                         }
                         else
                         {
@@ -324,25 +324,12 @@ namespace Caliimperium
 
             text += new Caliimperium.Temperatura().PegarTemperatura(cod)+"+";
             text += new Caliimperium.Temperatura().PegarTempMinima(cod) + "+";
-            text += new Caliimperium.Temperatura().Pegar1Quartil(cod) + "+";
+            text += new Caliimperium.Temperatura().Pegar2Quartil(cod) + "+";
             text += new Caliimperium.Temperatura().PegarMediana(cod) + "+";
             text += new Caliimperium.Temperatura().Pegar3Quartil(cod) + "+";
             text += new Caliimperium.Temperatura().PegarTempMaxima(cod) + "+";
             text += new Temperatura().PegarMinima(cod) + "+";
             text += new Temperatura().PegarMaxima(cod);
-
-            string[] textSplit = text.Split('+');
-
-            int c = 0;
-            while (c < textSplit.Length)
-            {
-                if (textSplit[c] == "")
-                {
-                    return "";
-                }
-                c++;
-            }
-
 
             return text;
         }
