+++ Cali_Imperium/Cali_Imperium/Sistema/css/padroes.css
@@ -315,9 +315,6 @@ div.sem_termometro{
     color: #fff;
 }
 
-div.fundo_modal_faq_on {
-
-
 /*inicio CSS tela de temperaturas ideais*/
 
 body {
@@ -329,7 +326,6 @@ div.modal_off {
 }
 
 div.modal_on {
-
     transition: all 4.3s;
     position: fixed;
     width: 100%;
@@ -342,75 +338,6 @@ div.modal_on {
     align-items: center;
     z-index: 5 !important;
 }
-
-div.fundo_modal_faq_off {
-    display: none;
-}
-div#faq {
-    position: fixed;
-    width: 700px;
-    height: 450px;
-    background-color: #fff;
-    padding-bottom: 30px;
-    padding-top: 30px;
-    padding-left: 30px;
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    /*border: solid 2px #000;*/
-    /*z-index: 9 !important; */
-}
-   .fundo_faq {
-        position: relative;
-        width: 100%;
-        height: 90%;
-        overflow-x: hidden;
-        overflow-y: scroll;
-    }
-.abrir {
-    position: relative;
-    width: 615px;
-    height: 50px;
-    background: #0171bb;
-    overflow: hidden;
-    transition: all 3.5s;
-    margin: 5px;
-}
-
-.abrir:hover {
-    height: auto;
-    transition: all 0.5s;
-}
-
-    .abrir:hover .resposta {
-        position: relative;
-        width: 100%;
-        height: auto;
-        background: #fff;
-        transition: all 0.5s;
-    }
-.pergunta {
-    position: relative;
-    width: 400px;
-    height: 50px;
-    font-family: Helvetica, sans-serif;
-    font-size: 20px;
-    color: #fff;
-    padding: 15px;
-    transition: all 0.5s;
-}
-.resposta {
-    position: relative;
-    width: 100%;
-    height: 300px;
-    font-family: Helvetica, sans-serif;
-    font-size: 20px;
-    color: black;
-    padding: 15px;
-    transition: all 0.5s;
-}
-
-
 
 div#tela_ideais {
     position: fixed;
@@ -482,4 +409,3 @@ div#tela_ideais {
 
 
 /*fim CSS tela de temperaturas ideais*/
-
