--- Cali_Imperium/Cali_Imperium/Sistema/css/relatorio.css
@@ -1,108 +1,0 @@
-﻿div.area_relatorios {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    background-color: #fff;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-div.relatorios {
-    position: relative;
-    width: 700px;
-    height: 450px;
-    display: flex;
-    justify-content: center;
-    background-color: #ccc;
-    flex-wrap: wrap;
-}
-
-div.opcoes_relatorios {
-    position: relative;
-    width: 100%;
-    height: 40px;
-    background-color: #4332d4;
-    padding: 10px;
-    overflow: hidden;
-}
-
-a.opc_relat {
-    position: relative;
-    text-decoration: none;
-    color: #fff;
-    padding: 14px 5px;
-    /*background-color: #000;*/
-    font-family: 'Century Gothic', sans-serif;
-}
-
-table#data_minima.calendario, table#data_maxima.calendario {
-    margin: 15px;
-    width: 260px;
-    height: 150px;
-}
-
-.prox_mes {
-    transition: all 0.3s;
-    width: 100%;
-    text-align: center;
-    cursor: pointer;
-}
-
-.prox_mes a {
-   text-decoration: none;
-   font-family: "Century Gothic";
-   padding-left: 14px;
-   padding-right: 14px;
-}
-
-.dia_selecionado {
-    background: #4332d4 !important;
-}
-
-.prox_mes:hover {
-    background: #4332d4;
-    transition: all 0.3s;
-}
-
-td a{
-    text-decoration: none;
-}
-
-.calendario {
-    position: relative;
-    background-color: #fff;
-    font-family: "Century Gothic";
-}
-
-.nome_dias {
-    background: #4332d4;
-    color: #fff;
-    padding: 5px;
-}
-
-.dias_do_mes {
-    background: #00ff68;
-}
-
-input.btn {
-    width: 80%;
-    height: 40px;
-    background: #00dd31;
-    border: none;
-}
-
-div.modal_relatorio {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(72, 0, 255, 0.44);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    flex-wrap: wrap;
-    overflow: scroll;
-}
-p.numeros{
-    
-}
