+++ Cali_Imperium/Cali_Imperium/Sistema/js/eventos.js
@@ -214,19 +214,6 @@ function removerViewAddArduino() {// mesma coisa da removerViewConfiguracoes
     div_fundo.className = "fundo_modal_arduino_off";
 }
 
-function getViewFaq() { // mesma coisa da getViewConfiguracoes
-    var div_fundo = document.getElementById("fundo_modal_faq");
-    div_fundo.className = "";
-    div_fundo.className = "fundo_modal_faq_on";
-}
-
-function removerViewFaq() {// mesma coisa da removerViewConfiguracoes
-    var div_fundo = document.getElementById("fundo_modal_faq");
-    div_fundo.className = "";
-    div_fundo.className = "fundo_modal_faq_off";
-}
-
-
 /* abrir e fechar janela modal temperatura modal */
 function abrirmodal() {
     var div_fundo = document.getElementById("modal");
@@ -240,4 +227,3 @@ function fecharmodal() {
     div_fundo.className = "modal_off";
 }
 /* fim janela modal temperaturA */
-
