--- Cali_Imperium/tela_relatorio/index.html
@@ -1,202 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<title></title>
-	<style type="text/css">
-		*{
-			margin: 0px;
-			padding: 0px;
-			box-sizing: border-box;
-		}
-		.pagina_relatorio{
-			width: 598px;
-			min-height: 842px;
-			/*background: #ddd;*/
-			border-right-style: solid;
-		}
-		.cabeca{
-			position: absolute;
-			width: 595px;
-			height: 150px;
-			/*background: #8da978;*/
-		}
-		img{
-			width: 100px;
-			height: 100px;
-			margin: 20px;
-			background-size: 100% 100%
-		}
-		div.dados{
-			position: absolute;
-			width: 450px;
-			height: 150px;
-			/*background: #f0f;*/
-			right: 0px;
-			padding: 15px;
-
-		}
-		div.dados p{
-			color: #000;
-			font-family: "Century Gothic";
-			text-align: right;
-		}
-		div.dados p.email{
-			color: #000;
-			position: absolute;
-			font-family: "Century Gothic";
-			right: 15px;
-			bottom: 15px;
-		}
-		div.infos{
-			position: absolute;
-			width: 595px;
-			min-height: 692px;
-			/*background: #faa;*/
-			top: 150px;
-			padding: 10px;
-		}
-		table{
-			position: relative;
-			width: 100%;
-			text-align: center;	
-			color: #000;
-			font-family: "Century Gothic";	
-			border: solid;
-			border-width: 1px;
-		}
-		td{
-			border: solid;
-			padding: 10px;
-			border-width: 1px;
-		}
-	</style> 
-</head>
-<body>
-	<div class="pagina_relatorio">
-		<div class="cabeca">
-
-			<div class="dados">
-				<p>05/06/2018</p>
-				<p>Relatorio por periodo</p>
-				<p class="email">mateus.covos@gmail.com</p>
-			</div>
-			<img src="logo.png">
-		</div>
-
-		<div class="infos">
-			<table> 
-				<tr> 
-					<td>Menor temperatura</td>
-					<td>Temperatura média</td>
-					<td>Maior temperatura</td>
-				</tr>
-				<tr> 
-					<td id="txtMenorTemperatura">10ºc</td>
-					<td id="txtTemperaturaMedia">20ºc</td>
-					<td id="txtMaiorTemperatura">30ºc</td>
-				</tr>	
-			</table>
-				
-				<br> 
-				<br> 
-				<br> 
-
-			<table>		
-				<tr> 
-					<td>Data</td>
-					<td>Temperatura</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-				
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>	
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-
-				<tr> 
-					<td id="txtData">05/06/2018 - 12.23</td>
-					<td id="txtTemperatura">20ºc</td>
-				</tr>
-			</table>
-		</div>
-
-	</div>
-</body>
-</html>
-
-<script>
-window.onload = function(){
-
-		var tr = document.getElementsByTagName("tr");
-
-		for(var i = 0; i < tr.length; i++){
-
-			if(i % 2 == 0){
-
-				tr[i].style.backgroundColor = "#aaa";
-
-			}else{
-
-				tr[i].style.backgroundColor = "#ddd";	
-
-			}
-
-		}
-
-
-}
-</script>
